requirements.txt
accelerate==0.26.1
aiohttp==3.8.6
aiolimiter==1.1.0
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.6.0
anthropic==0.3.11
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
black==23.11.0
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
contourpy==1.2.0
cycler==0.12.1
dask==2023.11.0
decorator==5.1.1
distro==1.8.0
dnspython==2.4.2
fastapi==0.104.1
filelock==3.13.1
flake8==6.1.0
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
gcsfs==2023.10.0
google-api-core==2.13.0
google-auth==2.23.4
google-auth-oauthlib==1.2.0
google-cloud-core==2.3.3
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpclib==0.4.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.17.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
isort==5.12.0
jaraco.classes==3.3.0
Jinja2==3.1.2
joblib==1.3.2
jsonlines==4.0.0
keyring==24.2.0
keyrings.google-artifactregistry-auth==1.1.2
kiwisolver==1.4.5
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
mccabe==0.7.0
mdurl==0.1.2
modal==0.55.4073
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.1
oauthlib==3.2.2
packaging==23.2
pandas==2.1.2
partd==1.4.1
pathspec==0.11.2
Pillow==10.1.0
platformdirs==4.0.0
pluggy==1.3.0
protobuf==4.25.0
psutil==5.9.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pydantic==2.4.2
pydantic_core==2.10.1
pyflakes==3.1.0
Pygments==2.16.1
pymongo==4.6.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
replicate==0.15.8
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sigtools==4.0.1
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
swifter==1.4.0
sympy==1.12
synchronicity==0.5.3
tblib==3.0.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
tokencost==0.0.5
tokenizers==0.14.1
toml==0.10.2
toolz==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transformers==4.35.0
typer==0.9.0
types-certifi==2021.10.8.3
types-toml==0.10.8.7
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
watchfiles==0.21.0
yarl==1.9.2
zipp==3.17.0
jsonargparse==4.27.5
